+++ MK4duo/src/core/commands/gcode/delta/g33_type2.h
@@ -70,9 +70,6 @@ static void ac_setup(const bool reset_bed) {
     if (reset_bed) bedlevel.reset(); // After full calibration bed-level data is no longer valid
   #endif
 
-  #if HAS_SD_STATISTICS
-    printer.setStatisticsStore(false);
-  #endif
 }
 
 static void ac_cleanup() {
@@ -87,9 +84,6 @@ static void ac_cleanup() {
     tools.change(old_tool_index, 0, true);
   #endif
 
-  #if HAS_SD_STATISTICS
-    printer.setStatisticsStore(true);
-  #endif
 }
 
 static void Report_signed_float(PGM_P const prefix, const float &f) {