+++ MK4duo/src/core/tools/nozzle.h
@@ -19,9 +19,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-
-#ifndef __NOZZLE_H__
-#define __NOZZLE_H__
+#pragma once
 
 #if ENABLED(NOZZLE_CLEAN_FEATURE) || ENABLED(NOZZLE_PARK_FEATURE)
 
@@ -85,5 +83,3 @@
   };
 
 #endif // ENABLED(NOZZLE_CLEAN_FEATURE) || ENABLED(NOZZLE_PARK_FEATURE)
-
-#endif /* __NOZZLE_H__ */