+++ MK4duo/src/feature/caselight/sanitycheck.h
@@ -19,6 +19,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
+#pragma once
 
 /**
  * sanitycheck.h
@@ -26,22 +27,17 @@
  * Test configuration values for errors at compile-time.
  */
 
-#ifndef _CASELIGHT_SANITYCHECK_H_
-#define _CASELIGHT_SANITYCHECK_H_
-
-  #if ENABLED(CASE_LIGHT)
-    #if !PIN_EXISTS(CASE_LIGHT) && DISABLED(CASE_LIGHT_USE_NEOPIXEL)
-      #error "DEPENDENCY ERROR: You have to set CASE_LIGHT_PIN to a valid pin if you enable CASE_LIGHT."
-    #endif
-    #if DISABLED(INVERT_CASE_LIGHT)
-      #error "DEPENDENCY ERROR: Missing setting INVERT_CASE_LIGHT is needed by CASE_LIGHT."
-    #endif
-    #if DISABLED(CASE_LIGHT_DEFAULT_ON)
-      #error "DEPENDENCY ERROR: Missing setting CASE_LIGHT_DEFAULT_ON is needed by CASE_LIGHT."
-    #endif
-    #if DISABLED(CASE_LIGHT_DEFAULT_BRIGHTNESS)
-      #error "DEPENDENCY ERROR: Missing setting CASE_LIGHT_DEFAULT_BRIGHTNESS is needed by CASE_LIGHT."
-    #endif
-  #endif // ENABLED(CASE_LIGHT)
-
-#endif /* _CASELIGHT_SANITYCHECK_H_ */
+#if ENABLED(CASE_LIGHT)
+  #if !PIN_EXISTS(CASE_LIGHT) && DISABLED(CASE_LIGHT_USE_NEOPIXEL)
+    #error "DEPENDENCY ERROR: You have to set CASE_LIGHT_PIN to a valid pin if you enable CASE_LIGHT."
+  #endif
+  #if DISABLED(INVERT_CASE_LIGHT)
+    #error "DEPENDENCY ERROR: Missing setting INVERT_CASE_LIGHT is needed by CASE_LIGHT."
+  #endif
+  #if DISABLED(CASE_LIGHT_DEFAULT_ON)
+    #error "DEPENDENCY ERROR: Missing setting CASE_LIGHT_DEFAULT_ON is needed by CASE_LIGHT."
+  #endif
+  #if DISABLED(CASE_LIGHT_DEFAULT_BRIGHTNESS)
+    #error "DEPENDENCY ERROR: Missing setting CASE_LIGHT_DEFAULT_BRIGHTNESS is needed by CASE_LIGHT."
+  #endif
+#endif