+++ MK4duo/src/lcd/ultralcd/dogm/status_screen_dogm.cpp
@@ -402,7 +402,7 @@ void LcdUI::draw_status_screen() {
         lcd_put_u8str(buffer2);
       }
       else {
-        lcd_put_u8str(itostr4(powerManager.consumption_hour - powerManager.startpower));
+        lcd_put_u8str(itostr4(print_job_counter.getConsumptionHour() - powerManager.startpower));
         lcd_put_u8str((char*)"Wh");
       }
     #else
@@ -522,7 +522,7 @@ void LcdUI::draw_status_screen() {
             lcd_put_u8str_P(PSTR("P:"));
             lcd_put_u8str(ftostr31(powerManager.consumption_meas));
             lcd_put_u8str_P(PSTR("W C:"));
-            lcd_put_u8str(ltostr7(powerManager.consumption_hour));
+            lcd_put_u8str(ltostr7(print_job_counter.getConsumptionHour()));
             lcd_put_u8str_P(PSTR("Wh"));
           }
       #endif